+++ packages/core-js-compat/src/modules-by-versions.js
@@ -13,6 +13,7 @@ module.exports = {
   ],
   3.3: [
     'esnext.async-iterator.constructor',
+    'esnext.async-iterator.as-indexed-pairs',
     'esnext.async-iterator.drop',
     'esnext.async-iterator.filter',
     'esnext.async-iterator.for-each',