+++ tests/commonjs.js
@@ -363,6 +363,7 @@ for (const _PATH of ['../packages/core-js-pure', '../packages/core-js']) {
   ok(typeof load('features/composite-key')({}, 1, {}) === 'object');
   ok(typeof load('features/composite-symbol')({}, 1, {}) === 'symbol');
   ok(typeof load('features/async-iterator') === 'function');
+  ok(typeof load('features/async-iterator/as-indexed-pairs') === 'function');
   ok(typeof load('features/async-iterator/drop') === 'function');
   ok(typeof load('features/async-iterator/filter') === 'function');
   ok(typeof load('features/async-iterator/for-each') === 'function');