+++ tests/pure/index.js
@@ -142,6 +142,7 @@ QUnit.module('ESNext');
 import './esnext.aggregate-error';
 import './esnext.array.is-template-object';
 import './esnext.async-iterator.constructor';
+import './esnext.async-iterator.as-indexed-pairs';
 import './esnext.async-iterator.drop';
 import './esnext.async-iterator.filter';
 import './esnext.async-iterator.for-each';