+++ tests/tests/index.js
@@ -193,6 +193,7 @@ import './esnext.array.is-template-object';
 import './esnext.array.last-item';
 import './esnext.array.last-index';
 import './esnext.async-iterator.constructor';
+import './esnext.async-iterator.as-indexed-pairs';
 import './esnext.async-iterator.drop';
 import './esnext.async-iterator.filter';
 import './esnext.async-iterator.for-each';